+++ src/components/Button/Button.tsx
@@ -9,11 +9,11 @@ import bem from "../../utils/bem";
 import { ButtonTypes } from "./ButtonTypes";
 import Icon from "../Icons/Icon";
 
+type ButtonElementType = "submit" | "button" | "reset";
+
 interface ButtonProps {
   /** Additional attributes passed to the button */
   attributes?: { [key: string]: any };
-  /** BlockName for use with BEM. See how to work with blockNames and BEM here: http://getbem.com/introduction/ */
-  blockName?: string;
   /** The kind of button assigned through the `ButtonTypes` enum  */
   buttonType?: ButtonTypes;
   /** Additional className for use with BEM. See how to work with blockNames and BEM here: http://getbem.com/introduction/ */
@@ -22,48 +22,47 @@ interface ButtonProps {
   disabled?: boolean;
   /** ID that other components can cross reference for accessibility purposes */
   id?: string;
-  /** Modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
-  modifiers?: string[];
   /** Trigger the Button's action through the `mouseDown` event handler instead of `onClick`. `false` by default. */
   mouseDown?: boolean;
   /** The action to perform on the `<button>`'s onClick function */
   onClick?: (event: React.MouseEvent | React.KeyboardEvent) => void;
   /** The html button attribute */
-  type?: "submit" | "button" | "reset";
+  type?: ButtonElementType;
 }
 
+// Used to map between ButtonTypes enum values and Chakra variant options.
+const variantMap = {};
+for (const type in ButtonTypes) {
+  variantMap[ButtonTypes[type]] = ButtonTypes[type];
+}
+
+/**
+ * Map the ButtonType to the Button Chakra theme variant object. If a wrong
+ * value is passed (typically in non-Typescript scenarios), then the default
+ * is the "primary" variant.
+ */
+const getVariant = (buttonType) =>
+  variantMap[buttonType] || ButtonTypes.Primary;
+
 /** Renders a simple `button` element with custom classes and modifiers. */
-// const Button = (props: React.PropsWithChildren<ButtonProps>) => {
 function Button(props: React.PropsWithChildren<ButtonProps>) {
   const {
     attributes,
-    blockName,
     buttonType,
     children,
     className = "",
     disabled,
     id,
-    modifiers = [],
     mouseDown = false,
     onClick,
-    type = "submit",
+    type = "button",
   } = props;
-
-  const buttonModifiers = modifiers ? modifiers : [];
-  if (
-    buttonType &&
-    !buttonModifiers.find((modifier) => modifier === buttonType)
-  ) {
-    buttonModifiers.push(buttonType);
-  }
   const baseClass = "button";
-  const btnCallback = mouseDown
-    ? { onMouseDown: onClick }
-    : { onClick: onClick };
-
+  const btnCallback = mouseDown ? { onMouseDown: onClick } : { onClick };
   let childCount = 0;
   let hasIcon = false;
   let variant;
+  let styles;
 
   React.Children.map(children, (child: React.ReactElement) => {
     childCount++;
@@ -78,32 +77,17 @@ function Button(props: React.PropsWithChildren<ButtonProps>) {
   });
 
   if (childCount === 1 && hasIcon) {
-    buttonModifiers.push("icon-only");
     variant = "icon-only";
   } else {
-    switch (buttonType) {
-      case ButtonTypes.Link:
-        variant = "link";
-        break;
-      case ButtonTypes.Pill:
-        variant = "pill";
-        break;
-      case ButtonTypes.Secondary:
-        variant = "secondary";
-        break;
-      case ButtonTypes.Primary:
-      default:
-        variant = "primary";
-        break;
-    }
+    variant = getVariant(buttonType);
   }
 
-  const styles = useStyleConfig("Button", { variant });
+  styles = useStyleConfig("Button", { variant });
 
   return (
     <ChakraButton
       id={id}
-      className={bem(baseClass, buttonModifiers, blockName, [className])}
+      className={bem(baseClass, [], "", [className])}
       type={type}
       disabled={disabled}
       __css={styles}